Add new "Mock" Classes
