--- app/src/main/java/com/spbu/allergy/UpdateableFragment.kt
@@ -1,5 +1,0 @@
-package com.spbu.allergy
-
-interface UpdatableFragment {
-    fun Update(someData : Data)
-}
