+++ app/src/main/java/com/spbu/allergy/buisnessLogic/Data.kt
@@ -1,0 +1,3 @@
+package com.spbu.allergy.buisnessLogic
+
+class Data
