+++ app/src/main/java/com/spbu/allergy/buisnessLogic/SomeApi.kt
@@ -1,0 +1,7 @@
+package com.spbu.allergy.buisnessLogic
+
+class SomeApi() {
+    fun request(user : UserData): Data {
+        return Data()
+    }
+}
