+++ app/src/main/java/com/spbu/allergy/buisnessLogic/UpdateableFragment.kt
@@ -1,0 +1,5 @@
+package com.spbu.allergy.buisnessLogic
+
+interface UpdatableFragment {
+    fun update(currentData : Data)
+}
