+++ app/src/main/java/com/spbu/allergy/buisnessLogic/UserData.kt
@@ -1,0 +1,3 @@
+package com.spbu.allergy.buisnessLogic
+
+class UserData{}
