+++ app/src/main/java/com/spbu/allergy/buisnessLogic/UserProfile.kt
@@ -1,0 +1,7 @@
+package com.spbu.allergy.buisnessLogic
+
+class UserProfile{
+    fun getInfo(): UserData {
+        return UserData()
+    }
+}
