+++ app/src/main/java/com/spbu/allergy/map/MapFragment.kt
@@ -15,10 +15,11 @@ import com.google.android.gms.maps.model.CameraPosition
 import com.google.android.gms.maps.model.LatLng
 import com.spbu.allergy.R
 import com.spbu.allergy.R.layout.fragment_map
-import com.spbu.allergy.UpdateableFragment
+import com.spbu.allergy.buisnessLogic.Data
+import com.spbu.allergy.buisnessLogic.UpdatableFragment
 
 
-class MapFragment(hasGeolocationPermission:()->Boolean) : Fragment(), UpdateableFragment{
+class MapFragment(hasGeolocationPermission:()->Boolean) : Fragment(), UpdatableFragment{
 
     private var geolocationPermissions = hasGeolocationPermission
     private lateinit var mMapView: MapView
@@ -94,7 +95,7 @@ class MapFragment(hasGeolocationPermission:()->Boolean) : Fragment(), Updateable
         mMapView.onLowMemory()
     }
 
-    override fun Update() {
+    override fun update(currentData : Data) {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 }
